+++ packages/editor-ui/src/components/NodeCreator/CategoryItem.vue
@@ -1,7 +1,7 @@
 <template>
 	<div :class="$style.category">
 		<span :class="$style.name">
-			{{ renderCategoryName(categoryName) }}
+			{{ renderCategoryName(categoryName) }} ({{ nodesCount }})
 		</span>
 		<font-awesome-icon
 			:class="$style.arrow"
@@ -16,12 +16,44 @@
 import Vue from 'vue';
 import camelcase from 'lodash.camelcase';
 import { CategoryName } from '@/plugins/i18n';
+import { INodeCreateElement, ICategoriesWithNodes } from '@/Interface';
+import { REGULAR_NODE_FILTER, TRIGGER_NODE_FILTER, ALL_NODE_FILTER } from '@/constants';
+
+const nodeTypeCountMapper = {
+	[REGULAR_NODE_FILTER]: ['regularCount'],
+	[TRIGGER_NODE_FILTER]: ['triggerCount'],
+	[ALL_NODE_FILTER]: ['triggerCount', 'regularCount'],
+};
 
 export default Vue.extend({
 	props: ['item'],
 	computed: {
+		selectedType(): "Regular" | "Trigger" | "All" {
+			return this.$store.getters['ui/selectedNodeCreatorType'];
+		},
+		categoriesWithNodes(): ICategoriesWithNodes {
+			return this.$store.getters['nodeTypes/categoriesWithNodes'];
+		},
+		categorizedItems(): INodeCreateElement[] {
+			return this.$store.getters['nodeTypes/categorizedItems'];
+		},
 		categoryName() {
 			return camelcase(this.item.category);
+		},
+		nodesCount(): number {
+			const currentCategory = this.categoriesWithNodes[this.item.category];
+			const subcategories = Object.keys(currentCategory);
+
+			const count = subcategories.reduce((accu: number, subcategory: string) => {
+				const countKeys = nodeTypeCountMapper[this.selectedType];
+
+				for (const countKey of countKeys) {
+					accu += currentCategory[subcategory][(countKey as "triggerCount" | "regularCount")];
+				}
+
+				return accu;
+			}, 0);
+			return count;
 		},
 	},
 	methods: {
